update int32->int64 将鉴权中间件添加

+++ idl/comment.thrift
@@ -3,15 +3,15 @@ namespace go comment
 include "user.thrift"
 
 struct Comment {
-    1: i32 id, // 视频评论id
+    1: i64 id, // 视频评论id
     2: user.User user, // 评论用户信息
     3: string content, // 评论内容
     4: string create_date, // 评论发布日期，格式 mm-dd
 }
 
 struct CommentActionRequest {
-    1: i32 user_id, // 用户鉴权token
-    2: i32 video_id, // 视频id
+    1: i64 user_id, // 用户鉴权token
+    2: i64 video_id, // 视频id
     3: i32 action_type, // 1-发布评论，2-删除评论
     4: optional string comment_text, // 用户填写的评论内容，在action_type=1的时候使用
     5: optional i32 comment_id, // 要删除的评论id，在action_type=2的时候使用
@@ -24,8 +24,8 @@ struct CommentActionResponse {
 }
 
 struct CommentListRequest {
-    1: i32 user_id, // 用户id
-    2: i32 video_id, // 视频id
+    1: i64 user_id, // 用户id
+    2: i64 video_id, // 视频id
 }
 
 struct CommentListResponse {
@@ -37,5 +37,5 @@ struct CommentListResponse {
 service CommentService {
     CommentActionResponse CommentAction(1: CommentActionRequest Request),
     CommentListResponse GetCommentList(1: CommentListRequest Request),
-    i32 GetCommentCount(1: i32 video_id), //根据video_id获取评论数
+    i32 GetCommentCount(1: i64 video_id), //根据video_id获取评论数
 }
